+++ datafixtures/__init__.py
@@ -1,4 +1,4 @@
 
 
 VERSION = "1.0.0"
-"""Application version number tuple."""
+"""Application version number."""